dom/xslt/tests/mochitest/test_bug551654.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=551654
-->
<head>
  <title>Test for Bug 551654</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=551654">Mozilla Bug 551654</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 551654 **/

  var didTransform = false;
  var processor = new XSLTProcessor();
  var style =
    '<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" ' +
                    'xmlns:exsl="http://exslt.org/common" ' +
                    'version="1.0">' +
      '<xsl:output method="html"/>' +
      '<xsl:template match="/">' +
        '<xsl:copy-of select="exsl:node-set(42)"/>' +
      '</xsl:template>' +
    '</xsl:stylesheet>';
  var styleDoc = new DOMParser().parseFromString (style, "text/xml");

  var data =
    '<root/>';
  var originalDoc = new DOMParser().parseFromString(data, "text/xml");

  processor.importStylesheet(styleDoc);
  var fragment = processor.transformToFragment(originalDoc, document);
  is(fragment.firstChild.nodeType, Node.TEXT_NODE,
     "Result of transform should be a textnode");
  is(fragment.firstChild.nodeValue, "42",
     "Result of transform should be a textnode with value '42'");

</script>
</pre>
</body>
</html>
